requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0

# ML and NLP
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
torch>=2.0.0
transformers>=4.30.0
langdetect==1.0.9

# PDF Processing
pdfplumber==0.10.3
pdfminer.six==20221105
pandas==2.0.3

# Translation (optional)
# googletrans==4.0.0rc1
# indic-nlp-library==0.92

# Security and middleware
slowapi==0.1.9
prometheus-fastapi-instrumentator==6.1.0

# Data processing
orjson==3.9.10
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.0
